mock django-webpack-loader in view tests
django-webpack-loader expects there to always
be the `webpack-stats.json` file. When we are testing
in a clean environment (such as Travis CI) that
file will not be there. We can get around this
by mocking the template tag function completely
so that the code that looks for the file is never called