Manage django-webpack-loader in local_settings.py
This was set to `not DEBUG` but `DEBUG` was only set later
in the local_settings.py so this would always evaluate to
`True`. It makes sense to make this explicit. To actually
disable the cache when developing the following line should be added
to `local_settings.py`:

```
WEBPACK_LOADER["DEFAULT"]["CACHE"] = False
```